Move rest code to rest package and add constructor function
